backend/AiService/backend/requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
numpy>=1.21.0,<2.0.0
tensorflow>=2.17.0
Pillow>=10.1.0
opencv-python>=4.8.0
